Allow getUser override
