+++ .changeset/long-dryers-clap.md
@@ -1,0 +1,6 @@
+---
+'@roll-network/session-manager': minor
+'@roll-network/auth-sdk': minor
+---
+
+Allow getUser API function override
