+++ packages/auth-sdk/src/types/config.ts
@@ -1,0 +1,9 @@
+export interface Config {
+  clientId: string
+  issuerUrl: string
+  redirectUrl: string
+  logoutRedirectUrl: string
+  scopes: string[]
+  apiUrl: string
+  clientSecret?: string
+}
