+++ packages/auth-sdk/src/types/event.ts
@@ -1,0 +1,4 @@
+export enum Event {
+  CredentialsUpdated = 'credentialsrupdated',
+  CredentialsCreated = 'credentialscreated',
+}
