+++ packages/auth-sdk/src/types/override.ts
@@ -1,0 +1,6 @@
+import { User } from './credentials'
+import { Token } from './token'
+
+export interface Override {
+  getUser?: <U extends User>(token: Token) => Promise<U>
+}
