+++ packages/session-manager/src/hooks/use-auth-sdk-user.ts
@@ -1,12 +1,12 @@
 import SDK, { Credentials, Event, User } from '@roll-network/auth-sdk'
 import { useEffect, useRef, useState } from 'react'
 
-const useAuthSdkUser = (authSdk: SDK) => {
+const useAuthSdkUser = <U extends User>(authSdk: SDK) => {
   const isSyncedRef = useRef(false)
-  const [user, setUser] = useState<User>()
+  const [user, setUser] = useState<U>()
 
   useEffect(() => {
-    const handleCredentialsChange = (credentials?: Credentials) => {
+    const handleCredentialsChange = (credentials?: Credentials<U>) => {
       if (credentials?.user) {
         setUser(credentials.user)
       } else {
