+++ packages/session-manager/src/types.ts
@@ -3,8 +3,8 @@ import SDK, { User } from '@roll-network/auth-sdk'
 
 export type SessionStatus = 'initializing' | 'refreshing' | 'stale'
 
-export interface SessionContextValue {
-  user?: User
+export interface SessionContextValue<U extends User = User> {
+  user?: U
   logIn: () => Promise<void>
   logOut: () => Promise<void>
   refresh: () => Promise<void>
